requirements.txt
PySide6>=6.7
polars>=1.10
pyarrow>=16.0
openpyxl>=3.1
python-docx>=1.1
oracledb>=2.4
python-dotenv>=1.0
pandas>=2.2
